added product options

+++ backend/src/services/product.service.ts
@@ -1,6 +1,7 @@
 import { ImageUploadHelper } from '../helpers/image-upload';
 import { validateFile } from '../middleware/validation/file';
 import { ProductModel } from '../models/products.model';
+import { WhatSappSettingsModel } from '../models/whatsapp-settings.model';
 import { Pagination } from '../types/common-types';
 import { File } from '../types/file';
 import { IProduct } from '../types/product';
@@ -9,6 +10,8 @@ import { User } from '../types/users';
 export class ProductService {
   static async createProduct(product: IProduct, user: Pick<User, 'id' | 'organizationId'>, file: File) {
     if (!user.organizationId) throw new Error('kindly create an organization to continue');
+    const whatsappData = await WhatSappSettingsModel.findOne({ where: { organizationId: user.organizationId } });
+    if (!whatsappData?.catalogId) throw new Error("you don't have an active catalog, kindly create one");
     const { valid, errors } = validateFile(file);
     if (!valid) throw new Error(errors.join(', '));
     const manageImageFile = new ImageUploadHelper();
@@ -23,14 +26,19 @@ export class ProductService {
       { where: { id: createdProduct.id }, returning: true }
     );
   }
+  
   static async updateProduct(product: IProduct, user: Pick<User, 'id' | 'organizationId'>, file: File) {
     const { id, ...productWithOutId } = product;
     if (!id) throw new Error('product id is required');
     if (!user.organizationId) throw new Error('kindly create an organization to continue');
+    const whatsappData = await WhatSappSettingsModel.findOne({ where: { organizationId: user.organizationId } });
+    if (!whatsappData?.catalogId) throw new Error("you don't have an active catalog, kindly create one");
     return await ProductModel.update(productWithOutId, { where: { id: id }, returning: true });
   }
   static async removeProduct(productId: string, user: Pick<User, 'id' | 'organizationId'>) {
     if (!productId) throw new Error('product id is required');
+    const whatsappData = await WhatSappSettingsModel.findOne({ where: { organizationId: user.organizationId } });
+    if (!whatsappData?.catalogId) throw new Error("you don't have an active catalog, kindly create one");
     return await ProductModel.destroy({ where: { id: productId } });
   }
   static async getProduct(productId: string, user: Pick<User, 'id' | 'organizationId'>) {
